requirements.txt
coverage==4.1
py==1.4.31
pypng==0.0.18
pytest==2.9.2
wheel==0.24.0
